requirements.txt
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.12.0
astroid==3.3.11
attrs==25.4.0
authlib==1.6.6
babel==2.17.0
bandit==1.7.9
bidict==0.23.1
black==25.12.0
blinker==1.9.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
configargparse==1.7.1
coverage==7.13.0
cryptography==46.0.3
dill==0.4.0
docutils==0.21.2
dparse==0.6.4
execnet==2.1.2
factory-boy==3.3.3
faker==39.0.0
filelock==3.12.4
flake8==7.1.1
flask==3.1.2
flask-cors==6.0.2
flask-login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.7
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imagesize==1.4.1
iniconfig==2.3.0
isort==5.13.2
itsdangerous==2.2.0
jinja2==3.1.6
locust==2.42.6
locust-cloud==1.30.0
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==4.1.1
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.2
mypy==1.13.0
mypy-extensions==1.1.0
outcome==1.3.0.post0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
playwright==1.57.0
pluggy==1.6.0
psutil==6.0.0
pycodestyle==2.12.1
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pyee==13.0.0
pyflakes==3.2.0
pygments==2.19.2
pylint==3.3.1
pysocks==1.7.1
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==6.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.14.0
pytest-timeout==2.3.1
pytest-xdist==3.6.1
python-dotenv==1.0.1
python-engineio==4.12.3
python-socketio==5.15.1
pytokens==0.3.0
pyyaml==6.0.2
pyzmq==27.1.0
requests==2.32.3
responses==0.25.3
rich==14.2.0
ruamel-yaml==0.18.17
ruamel-yaml-clib==0.2.15
safety==3.2.7
safety-schemas==0.0.17
selenium==4.39.0
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stevedore==5.6.0
tomlkit==0.13.3
trio==0.32.0
trio-websocket==0.12.2
typer==0.20.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.2
websocket-client==1.9.0
werkzeug==3.1.4
wsproto==1.3.2
zope-event==6.1
zope-interface==8.1.1
